usb: message: Fixed error: 'no space before bracket'
This patch fixes error: 'no space before bracket' in Usb/message.c

+++ drivers/usb/core/message.c
@@ -303,10 +303,9 @@ static void sg_complete(struct urb *urb)
 		 */
 		spin_unlock(&io->lock);
 		for (i = 0, found = 0; i < io->entries; i++) {
-			if (!io->urbs[i])
+			if (!io->urbs[i] || !io->urbs[i]->dev)
 				continue;
 			if (found) {
-				usb_block_urb(io->urbs[i]);
 				retval = usb_unlink_urb(io->urbs[i]);
 				if (retval != -EINPROGRESS &&
 				    retval != -ENODEV &&
@@ -517,6 +516,11 @@ void usb_sg_wait(struct usb_sg_request *io)
 		int retval;
 
 		io->urbs[i]->dev = io->dev;
+		retval = usb_submit_urb(io->urbs[i], GFP_ATOMIC);
+
+		/* after we submit, let completions or cancelations fire;
+		 * we handshake using io->status.
+		 */
 		spin_unlock_irq(&io->lock);
 
 		retval = usb_submit_urb(io->urbs[i], GFP_NOIO);
@@ -593,13 +597,22 @@ void usb_sg_cancel(struct usb_sg_request *io)
 	for (i = io->entries - 1; i >= 0; --i) {
 		usb_block_urb(io->urbs[i]);
 
-		retval = usb_unlink_urb(io->urbs[i]);
-		if (retval != -EINPROGRESS
-		    && retval != -ENODEV
-		    && retval != -EBUSY
-		    && retval != -EIDRM)
-			dev_warn(&io->dev->dev, "%s, unlink --> %d\n",
-				 __func__, retval);
+		io->status = -ECONNRESET;
+		spin_unlock(&io->lock);
+		for (i = 0; i < io->entries; i++) {
+			int retval;
+
+			if (!io->urbs[i]->dev)
+				continue;
+			retval = usb_unlink_urb(io->urbs[i]);
+			if (retval != -EINPROGRESS
+					&& retval != -ENODEV
+					&& retval != -EBUSY
+					&& retval != -EIDRM)
+				dev_warn(&io->dev->dev, "%s, unlink --> %d\n",
+					__func__, retval);
+		}
+		spin_lock(&io->lock);
 	}
 
 	spin_lock_irqsave(&io->lock, flags);
